Move mongo_helper.py to datastore/; update import paths
